Intermediate commit to use multiprocessing.
This commit refactors the code so now we have a function that operates on a host
to update all the domains for that host. The function returns a list of hosts so
that the caller can then proceed with the cleanup. This function also creates a log
for each host, this will save me the hassle of writing to the same log file when multiprocessing
is used.

The update_instance function is now called get_instance_metrics.
It now returns the metrics instead of sending it.
The nics and disks discovery packet is in the beginning of the metrics, so it's always
sent before the stats for those items are discovered.