Use the cpu usage timestamp when we send the metric
This way zabbix's dependent method for calculating the percentage
cpu usage will use the exact timestamp when the absolute cpu time was
collected; which means accurate cpu usage.

And we don't have to send the timestamp as a separate item.

We are making this change because a calculated item will always calculate,
regardless of if a new value was available or not. A dependent value, however,
is only calcuated once new items are available. We use preprocessing (change per second)
and some javascript (to divide the result by 10^7 and cap the limits) to get
the percentage.

Another point to note is that, you can't pass any other item (eg; the custom timestamp
we were storing) to dependent items. It uses the timestamp which is part of the metric.